+++ BALSAMIC/commands/config/case.py
@@ -1,8 +1,7 @@
 """Balsamic config case CLI."""
-import json
 import logging
-import os
 from pathlib import Path
+from typing import Dict
 
 import click
 
@@ -20,7 +19,6 @@ from BALSAMIC.commands.options import (
     OPTION_CASE_ID,
     OPTION_CLINICAL_SNV_OBSERVATIONS,
     OPTION_CLINICAL_SV_OBSERVATIONS,
-    OPTION_CONTAINER_VERSION,
     OPTION_FASTQ_PATH,
     OPTION_GENDER,
     OPTION_NORMAL_SAMPLE_NAME,
@@ -32,9 +30,11 @@ from BALSAMIC.commands.options import (
     OPTION_TUMOR_SAMPLE_NAME,
     OPTION_UMI,
     OPTION_UMI_TRIM_LENGTH,
+    OPTION_CACHE_VERSION,
 )
 from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV, Gender, AnalysisWorkflow
 from BALSAMIC.constants.cache import GenomeVersion
+from BALSAMIC.constants.constants import FileType
 from BALSAMIC.constants.paths import CONTAINERS_DIR
 from BALSAMIC.constants.workflow_params import VCF_DICT
 from BALSAMIC.models.analysis import ConfigModel
@@ -45,7 +45,8 @@ from BALSAMIC.utils.cli import (
     generate_graph,
     get_analysis_fastq_files_directory,
 )
-from BALSAMIC.utils.io import write_json
+from BALSAMIC.utils.io import write_json, read_json
+from BALSAMIC.utils.utils import get_absolute_paths_dict
 
 LOG = logging.getLogger(__name__)
 
@@ -56,6 +57,7 @@ LOG = logging.getLogger(__name__)
 @OPTION_ANALYSIS_WORKFLOW
 @OPTION_BACKGROUND_VARIANTS
 @OPTION_BALSAMIC_CACHE
+@OPTION_CACHE_VERSION
 @OPTION_CADD_ANNOTATIONS
 @OPTION_CANCER_GERMLINE_SNV_OBSERVATIONS
 @OPTION_CANCER_SOMATIC_SNV_OBSERVATIONS
@@ -63,7 +65,6 @@ LOG = logging.getLogger(__name__)
 @OPTION_CASE_ID
 @OPTION_CLINICAL_SNV_OBSERVATIONS
 @OPTION_CLINICAL_SV_OBSERVATIONS
-@OPTION_CONTAINER_VERSION
 @OPTION_FASTQ_PATH
 @OPTION_GENDER
 @OPTION_GENOME_VERSION
@@ -84,6 +85,7 @@ def case_config(
     analysis_workflow: AnalysisWorkflow,
     background_variants: Path,
     balsamic_cache: Path,
+    cache_version: str,
     cadd_annotations: Path,
     cancer_germline_snv_observations: Path,
     cancer_somatic_snv_observations: Path,
@@ -91,7 +93,6 @@ def case_config(
     case_id: str,
     clinical_snv_observations: Path,
     clinical_sv_observations: Path,
-    container_version: str,
     fastq_path: Path,
     gender: Gender,
     genome_version: GenomeVersion,
@@ -105,21 +106,14 @@ def case_config(
     umi: bool,
     umi_trim_length: int,
 ):
-    if container_version:
-        balsamic_version = container_version
-
-    reference_config = os.path.join(
-        balsamic_cache, balsamic_version, genome_version, "reference.json"
+    references_path: Path = Path(balsamic_cache, cache_version, genome_version)
+    references: Dict[str, Path] = get_absolute_paths_dict(
+        base_path=references_path,
+        data=read_json(Path(references_path, f"reference.{FileType.JSON}").as_posix()),
     )
-    with open(reference_config, "r") as config_file:
-        reference_dict = json.load(config_file)
-
-    cadd_annotations_path = {
-        "cadd_annotations": cadd_annotations,
-    }
+    cadd_annotations_path = {"cadd_annotations": cadd_annotations}
     if cadd_annotations:
-        reference_dict.update(cadd_annotations_path)
-
+        references.update(cadd_annotations_path)
     variants_observations = {
         "clinical_snv_observations": clinical_snv_observations,
         "clinical_sv_observations": clinical_sv_observations,
@@ -129,7 +123,7 @@ def case_config(
         "swegen_snv_frequency": swegen_snv,
         "swegen_sv_frequency": swegen_sv,
     }
-    reference_dict.update(
+    references.update(
         {
             observations: path
             for observations, path in variants_observations.items()
@@ -157,9 +151,9 @@ def case_config(
             "sequencing_type": "targeted" if panel_bed else "wgs",
             "analysis_workflow": analysis_workflow,
         },
-        reference=reference_dict,
+        reference=references,
         singularity={
-            "image": os.path.join(balsamic_cache, balsamic_version, "containers")
+            "image": Path(balsamic_cache, cache_version, "containers").as_posix()
         },
         background_variants=background_variants,
         samples=get_sample_list(