+++ BALSAMIC/commands/config/pon.py
@@ -3,10 +3,10 @@ import json
 import logging
 import os
 from pathlib import Path
+from typing import Dict
 
 import click
 
-from BALSAMIC import __version__ as balsamic_version
 from BALSAMIC.commands.options import (
     OPTION_GENOME_VERSION,
     OPTION_ADAPTER_TRIM,
@@ -19,9 +19,11 @@ from BALSAMIC.commands.options import (
     OPTION_QUALITY_TRIM,
     OPTION_UMI,
     OPTION_UMI_TRIM_LENGTH,
+    OPTION_CACHE_VERSION,
 )
 from BALSAMIC.constants.analysis import BIOINFO_TOOL_ENV
 from BALSAMIC.constants.cache import GenomeVersion
+from BALSAMIC.constants.constants import FileType
 from BALSAMIC.constants.paths import CONTAINERS_DIR
 from BALSAMIC.models.analysis import ConfigModel
 from BALSAMIC.utils.cli import (
@@ -30,7 +32,8 @@ from BALSAMIC.utils.cli import (
     get_pon_sample_list,
     get_analysis_fastq_files_directory,
 )
-from BALSAMIC.utils.io import write_json
+from BALSAMIC.utils.io import write_json, read_json
+from BALSAMIC.utils.utils import get_absolute_paths_dict
 
 LOG = logging.getLogger(__name__)
 
@@ -39,6 +42,7 @@ LOG = logging.getLogger(__name__)
 @OPTION_ADAPTER_TRIM
 @OPTION_ANALYSIS_DIR
 @OPTION_BALSAMIC_CACHE
+@OPTION_CACHE_VERSION
 @OPTION_CASE_ID
 @OPTION_FASTQ_PATH
 @OPTION_GENOME_VERSION
@@ -53,6 +57,7 @@ def pon_config(
     adapter_trim: bool,
     analysis_dir: Path,
     balsamic_cache: Path,
+    cache_version: str,
     case_id: str,
     fastq_path: Path,
     genome_version: GenomeVersion,
@@ -62,16 +67,14 @@ def pon_config(
     umi_trim_length: bool,
     version: str,
 ):
-    reference_config = os.path.join(
-        balsamic_cache, balsamic_version, genome_version, "reference.json"
+    references_path: Path = Path(balsamic_cache, cache_version, genome_version)
+    references: Dict[str, Path] = get_absolute_paths_dict(
+        base_path=references_path,
+        data=read_json(Path(references_path, f"reference.{FileType.JSON}").as_posix()),
     )
-    with open(reference_config, "r") as config_file:
-        reference_dict = json.load(config_file)
-
     fastq_path: str = get_analysis_fastq_files_directory(
         case_dir=Path(analysis_dir, case_id).as_posix(), fastq_path=fastq_path
     )
-
     config_collection_dict = ConfigModel(
         QC={
             "adapter_trim": adapter_trim,
@@ -89,9 +92,9 @@ def pon_config(
             "sequencing_type": "targeted" if panel_bed else "wgs",
         },
         samples=get_pon_sample_list(fastq_path),
-        reference=reference_dict,
+        reference=references,
         singularity={
-            "image": os.path.join(balsamic_cache, balsamic_version, "containers")
+            "image": Path(balsamic_cache, cache_version, "containers").as_posix()
         },
         bioinfo_tools=BIOINFO_TOOL_ENV,
         bioinfo_tools_version=get_bioinfo_tools_version(