+++ BALSAMIC/models/metrics.py
@@ -2,7 +2,7 @@
 import logging
 from typing import Optional, Any, List
 
-from pydantic import BaseModel, validator
+from pydantic.v1 import BaseModel, validator
 
 from BALSAMIC.constants.metrics import VALID_OPS